Add forwarding capability
Now, it's possible to configure the embedded version of the software as forwarder by defining PROXY_CONNECT
and setting master IP address and master port. The events will be the forwarded per UDP to the master.

The master processes events from all sources, filters out duplicates and then forwards the events to the
hue bridge.

This can be used to improve coverage in the house - just put an embedded forwarder in problematic areas.